Split drizzle-practice-session-repository.test.ts by method group
