+++ src/adapters/repositories/drizzle-practice-session-repository-test-helpers.ts
@@ -1,0 +1,6 @@
+import { vi } from 'vitest';
+
+export function restoreDrizzlePracticeSessionRepositoryTestMocks() {
+  vi.useRealTimers();
+  vi.restoreAllMocks();
+}
